Reunify files after mongo removal Remove evidence that we split the gem and reduce the number of files.
