+++ lib/audited/audit.rb
@@ -1,60 +1,45 @@
+require 'set'
+
 module Audited
-  module Audit
-    def self.included(klass)
-      klass.extend(ClassMethods)
-      klass.setup_audit
-    end
+  # Audit saves the changes to ActiveRecord models.  It has the following attributes:
+  #
+  # * <tt>auditable</tt>: the ActiveRecord model that was changed
+  # * <tt>user</tt>: the user that performed the change; a string or an ActiveRecord model
+  # * <tt>action</tt>: one of create, update, or delete
+  # * <tt>audited_changes</tt>: a serialized hash of all the changes
+  # * <tt>comment</tt>: a comment set with the audit
+  # * <tt>version</tt>: the version of the model
+  # * <tt>request_uuid</tt>: a uuid based that allows audits from the same controller request
+  # * <tt>created_at</tt>: Time that the change was performed
+  #
+  class Audit < ::ActiveRecord::Base
+    include ActiveModel::Observing
 
-    module ClassMethods
-      def setup_audit
-        belongs_to :auditable,  polymorphic: true
-        belongs_to :user,       polymorphic: true
-        belongs_to :associated, polymorphic: true
+    belongs_to :auditable,  polymorphic: true
+    belongs_to :user,       polymorphic: true
+    belongs_to :associated, polymorphic: true
 
-        before_create :set_version_number, :set_audit_user, :set_request_uuid
+    before_create :set_version_number, :set_audit_user, :set_request_uuid
 
-        cattr_accessor :audited_class_names
-        self.audited_class_names = Set.new
-      end
+    cattr_accessor :audited_class_names
+    self.audited_class_names = Set.new
 
-      # Returns the list of classes that are being audited
-      def audited_classes
-        audited_class_names.map(&:constantize)
-      end
+    serialize :audited_changes
 
-      # All audits made during the block called will be recorded as made
-      # by +user+. This method is hopefully threadsafe, making it ideal
-      # for background operations that require audit information.
-      def as_user(user, &block)
-        Thread.current[:audited_user] = user
-        yield
-      ensure
-        Thread.current[:audited_user] = nil
-      end
+    scope :ascending,     ->{ reorder(version: :asc) }
+    scope :descending,    ->{ reorder(version: :desc)}
+    scope :creates,       ->{ where(action: 'create')}
+    scope :updates,       ->{ where(action: 'update')}
+    scope :destroys,      ->{ where(action: 'destroy')}
 
-      # @private
-      def reconstruct_attributes(audits)
-        attributes = {}
-        result = audits.collect do |audit|
-          attributes.merge!(audit.new_attributes).merge!(version: audit.version)
-          yield attributes if block_given?
-        end
-        block_given? ? result : attributes
-      end
-
-      # @private
-      def assign_revision_attributes(record, attributes)
-        attributes.each do |attr, val|
-          record = record.dup if record.frozen?
-
-          if record.respond_to?("#{attr}=")
-            record.attributes.key?(attr.to_s) ?
-              record[attr] = val :
-              record.send("#{attr}=", val)
-          end
-        end
-        record
-      end
+    scope :up_until,      ->(date_or_time){where("created_at <= ?", date_or_time) }
+    scope :from_version,  ->(version){where(['version >= ?', version]) }
+    scope :to_version,    ->(version){where(['version <= ?', version]) }
+    scope :auditable_finder, ->(auditable_id, auditable_type){where(auditable_id: auditable_id, auditable_type: auditable_type)}
+    # Return all audits older than the current one.
+    def ancestors
+      self.class.ascending.where(['auditable_id = ? and auditable_type = ? and version <= ?',
+        auditable_id, auditable_type, version])
     end
 
     # Return an instance of what the object looked like at this revision. If
@@ -83,6 +68,64 @@ module Audited
       end
     end
 
+    # Allows user to be set to either a string or an ActiveRecord object
+    # @private
+    def user_as_string=(user)
+      # reset both either way
+      self.user_as_model = self.username = nil
+      user.is_a?(::ActiveRecord::Base) ?
+        self.user_as_model = user :
+        self.username = user
+    end
+    alias_method :user_as_model=, :user=
+    alias_method :user=, :user_as_string=
+
+    # @private
+    def user_as_string
+      user_as_model || username
+    end
+    alias_method :user_as_model, :user
+    alias_method :user, :user_as_string
+
+    # Returns the list of classes that are being audited
+    def self.audited_classes
+      audited_class_names.map(&:constantize)
+    end
+
+    # All audits made during the block called will be recorded as made
+    # by +user+. This method is hopefully threadsafe, making it ideal
+    # for background operations that require audit information.
+    def self.as_user(user, &block)
+      Thread.current[:audited_user] = user
+      yield
+    ensure
+      Thread.current[:audited_user] = nil
+    end
+
+    # @private
+    def self.reconstruct_attributes(audits)
+      attributes = {}
+      result = audits.collect do |audit|
+        attributes.merge!(audit.new_attributes).merge!(version: audit.version)
+        yield attributes if block_given?
+      end
+      block_given? ? result : attributes
+    end
+
+    # @private
+    def self.assign_revision_attributes(record, attributes)
+      attributes.each do |attr, val|
+        record = record.dup if record.frozen?
+
+        if record.respond_to?("#{attr}=")
+          record.attributes.key?(attr.to_s) ?
+            record[attr] = val :
+            record.send("#{attr}=", val)
+        end
+      end
+      record
+    end
+
     private
 
     def set_version_number
@@ -102,4 +145,5 @@ module Audited
       self.request_uuid ||= SecureRandom.uuid
     end
   end
+
 end
