--- spec/audited/adapters/active_record/active_record_spec_helper.rb
@@ -1,4 +1,0 @@
-require 'spec_helper'
-require 'audited/adapters/active_record'
-require 'support/active_record/models'
-load "audited/sweeper.rb" # force to reload sweeper
