+++ spec/spec_helper.rb
@@ -5,6 +5,8 @@ require 'rails_app/config/environment'
 require 'rspec/rails'
 require 'audited'
 require 'audited_spec_helpers'
+require 'support/active_record/models'
+load "audited/sweeper.rb" # force to reload sweeper
 
 SPEC_ROOT = Pathname.new(File.expand_path('../', __FILE__))
 
@@ -12,8 +14,4 @@ Dir[SPEC_ROOT.join('support/*.rb')].each{|f| require f }
 
 RSpec.configure do |config|
   config.include AuditedSpecHelpers
-
-  config.before(:each, :adapter => :active_record) do
-    Audited.audit_class = Audited::Adapters::ActiveRecord::Audit
-  end
 end
